--- rfclint/rfclint/foo
@@ -1,87 +1,0 @@
-Checking for well-formness of 'Tests\spell-utf8.xml'
-Well-formness passes
-Skipping schema validation
-No XML fragments in sourcecode elements found.
-No ABNF to check
-xx - 0.50.3
-Use iso8859
-spell command = 'c:\Program Files (x86)\Aspell\bin\aspell.exe -a --encoding=iso8859-1'
-@(#) International Ispell Version 3.1.20 (but really Aspell 0.50.3)
-
-Pre Encode = ^ Write some “correct” and “brken” words that are ‘tagged with don’t be odd about things.  Other fun words are cen‘t which should be cen’t.  can cen cin con cun
-
-^ Write some  correct  and  brken  words that are  tagged with don t be odd about things.  Other fun words are cen t which should be cen t.  can cen cin con cun
-
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = & brken 24 28: broken, bracken, Bren, Brien, Brena, Brenn, brake, braking, broke, broking, brine, Bergen, Breena, Brown, barren, bergen, brawn, broker, brown, Braden, Bremen, braked, brakes, brazen
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = & cen 36 111: CERN, cent, Chen, Ce, en, Sen, Zen, sen, zen, CEO, Ben, CNN, Can, Ced, Con, Den, Gen, Jen, Ken, Len, Pen, can, con, den, fen, gen, hen, ken, men, pen, ten, wen, yen, cine, Xena, ctn
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = *
-spell out line = & cen 36 133: CERN, cent, Chen, Ce, en, Sen, Zen, sen, zen, CEO, Ben, CNN, Can, Ced, Con, Den, Gen, Jen, Ken, Len, Pen, can, con, den, fen, gen, hen, ken, men, pen, ten, wen, yen, cine, Xena, ctn
-spell out line = *
-spell out line = *
-spell out line = & cen 36 145: CERN, cent, Chen, Ce, en, Sen, Zen, sen, zen, CEO, Ben, CNN, Can, Ced, Con, Den, Gen, Jen, Ken, Len, Pen, can, con, den, fen, gen, hen, ken, men, pen, ten, wen, yen, cine, Xena, ctn
-spell out line = & cin 40 149: cine, Cain, Chin, Con, Vin, chin, coin, con, Ci, IN, In, in, Crin, sin, cir, CID, CNN, Can, Cid, Fin, Kin, Lin, Min, PIN, Win, bin, can, cit, din, fin, gin, kin, min, pin, tin, win, yin, Xian, CIA, ctn
-spell out line = *
-spell out line = & cun 36 157: cunt, cub, cum, Cu, cu, UN, CNN, Can, Con, Gun, Jun, can, con, gun, jun, ctn, cue, cur, Dun, Hun, Sun, bun, cud, cup, cut, dun, fun, mun, nun, pun, run, sun, tun, Cain, coin, cuing
-spell out line = 
-Tests\spell-utf8.xml:4: Misspelled word was found 'brken'
-  Write some “ correct” and “ brken” words that are ‘ tagged with don’ t be odd about things.  Other fun words are 
-  broken, bracken, Bren, Brien, Brena, Brenn, brake, braking, broke, broking, brine, Bergen, Breena, Brown, barren, bergen, brawn, broker, brown, Braden, Bremen, braked, brakes, brazen
-Tests\spell-utf8.xml:4: Misspelled word was found 'cen'
-  words that are ‘ tagged with don’ t be odd about things.  Other fun words are cen‘ t which should be cen’ t.  can cen cin con cun 
-  CERN, cent, Chen, Ce, en, Sen, Zen, sen, zen, CEO, Ben, CNN, Can, Ced, Con, Den, Gen, Jen, Ken, Len, Pen, can, con, den, fen, gen, hen, ken, men, pen, ten, wen, yen, cine, Xena, ctn
-Tests\spell-utf8.xml:4: Misspelled word was found 'cen'
-  don’ t be odd about things.  Other fun words are cen‘ t which should be cen’ t.  can cen cin con cun 
-  CERN, cent, Chen, Ce, en, Sen, Zen, sen, zen, CEO, Ben, CNN, Can, Ced, Con, Den, Gen, Jen, Ken, Len, Pen, can, con, den, fen, gen, hen, ken, men, pen, ten, wen, yen, cine, Xena, ctn
-Tests\spell-utf8.xml:4: Misspelled word was found 'cen'
-  odd about things.  Other fun words are cen‘ t which should be cen’ t.  can cen cin con cun 
-  CERN, cent, Chen, Ce, en, Sen, Zen, sen, zen, CEO, Ben, CNN, Can, Ced, Con, Den, Gen, Jen, Ken, Len, Pen, can, con, den, fen, gen, hen, ken, men, pen, ten, wen, yen, cine, Xena, ctn
-Tests\spell-utf8.xml:4: Misspelled word was found 'cin'
-  about things.  Other fun words are cen‘ t which should be cen’ t.  can cen cin con cun 
-  cine, Cain, Chin, Con, Vin, chin, coin, con, Ci, IN, In, in, Crin, sin, cir, CID, CNN, Can, Cid, Fin, Kin, Lin, Min, PIN, Win, bin, can, cit, din, fin, gin, kin, min, pin, tin, win, yin, Xian, CIA, ctn
-Tests\spell-utf8.xml:4: Misspelled word was found 'cun'
-  Other fun words are cen‘ t which should be cen’ t.  can cen cin con cun 
-  cunt, cub, cum, Cu, cu, UN, CNN, Can, Con, Gun, Jun, can, con, gun, jun, ctn, cue, cur, Dun, Hun, Sun, bun, cud, cup, cut, dun, fun, mun, nun, pun, run, sun, tun, Cain, coin, cuing
-Pre Encode = ^ Street names in German have the word Straße and the city Köln as one of my favorates.
-
-^ Street names in German have the word Straße and the city Köln as one of my favorates.
-
-Traceback (most recent call last):
-  File "run.py", line 291, in <module>
-    main()
-  File "run.py", line 277, in main
-    speller.processTree(xmlrfc.tree.getroot())
-  File "d:\projects\v3\rfceditor\rfclint\rfclint\spell.py", line 339, in processTree
-    self.processTree(node)
-  File "d:\projects\v3\rfceditor\rfclint\rfclint\spell.py", line 339, in processTree
-    self.processTree(node)
-  File "d:\projects\v3\rfceditor\rfclint\rfclint\spell.py", line 339, in processTree
-    self.processTree(node)
-  File "d:\projects\v3\rfceditor\rfclint\rfclint\spell.py", line 337, in processTree
-    self.checkTree(tree)
-  File "d:\projects\v3\rfceditor\rfclint\rfclint\spell.py", line 343, in checkTree
-    results = self.processLine(wordSet)
-  File "d:\projects\v3\rfceditor\rfclint\rfclint\spell.py", line 288, in processLine
-    self.stdin.write(newLine)
-UnicodeEncodeError: 'ascii' codec can't encode character u'\xdf' in position 43: ordinal not in range(128)
